[A11] refactor: RecommendColumn.vue (#12)
