+++ src/views/MainPage.vue
@@ -15,24 +15,20 @@
       <div class="recommendHeader mt-4">
         <h1>推薦跟隨</h1>
       </div>
-      <RecommendColumn
-        v-for="recommendUser in recommendUsers"
-        :key="recommendUser.id"
-        :initial-recommenduser="recommendUser"
-      />
+      <RecommendColumn />
     </div>
     
   </div>
 </template>
 
 <script>
-import NavBar from "../components/NavBar.vue";
-import RecommendColumn from "../components/RecommendColumn.vue";
-import NavpillHeader from "../components/NavpillHeader.vue";
-import WrittingTweet from "../components/WrittingTweet.vue";
-import TweetList from "../components/TweetList.vue";
+import NavBar from "../components/NavBar.vue"
+import RecommendColumn from "../components/RecommendColumn.vue"
+import NavpillHeader from "../components/NavpillHeader.vue"
+import WrittingTweet from "../components/WrittingTweet.vue"
+import TweetList from "../components/TweetList.vue"
 import tweetsAPI from './../apis/tweets'
-import usersAPI from './../apis/users'
+// import usersAPI from './../apis/users'
 import { Toast } from './../utils/helpers'
 import store from './../store'
 
@@ -55,36 +51,8 @@ export default {
   },
   created () {
     this.fetchTweets()
-    this.fetchRecommendUsers()
   },
   methods: {
-    async fetchRecommendUsers() {
-      try {
-        this.isLoading = true
-
-        const { data } = await usersAPI.getUserFollowings({userId: this.currentUser.id})
-        const userFollowings = data
-
-        const responseUsers = await usersAPI.getTopUsers()
-        this.recommendUsers = responseUsers.data.map( user => {
-          return (
-            {
-              ...user,
-              isFollowed: userFollowings.some(f => f.followingId === user.id)
-            }
-          )
-        })
-
-        this.isLoading = false
-      } catch (error) {
-        console.error(error)
-        this.isLoading = false
-        Toast.fire({
-          icon: 'error',
-          title: '無法取得 RecommendUsers 資料，請稍後再試'
-        })
-      }
-    },
     async fetchTweets () {
       try {
         this.isLoading = true
