+++ src/views/UserSelf.vue
@@ -9,16 +9,22 @@
         <router-view />
       </div>
     </div>
-    <RecommendColumn />
+
+    <div id="recommendColumn-container">
+      <div class="recommendHeader mt-4">
+        <h1>推薦跟隨</h1>
+      </div>
+      <RecommendColumn />
+    </div>
   </div>
 </template>
 
 <script>
-import NavBar from "../components/NavBar.vue";
-import RecommendColumn from "../components/RecommendColumn.vue";
-import NavpillHeader from "../components/NavpillHeader.vue";
-import UserProfile from "../components/UserProfile.vue";
-import NavpillUser from "../components/NavpillUser.vue";
+import NavBar from "../components/NavBar.vue"
+import RecommendColumn from "../components/RecommendColumn.vue"
+import NavpillHeader from "../components/NavpillHeader.vue"
+import UserProfile from "../components/UserProfile.vue"
+import NavpillUser from "../components/NavpillUser.vue"
 
 export default {
   name: "UserSelf",
